// glendy/src/app/models/producto.ts
export class Producto {
    _id :string;
    nombre_producto: String;
    tipo_producto: String;
    descripcion: String;
    contenido: String[];
    precio_venta: number;
    precio_costo: number;
    existencia: number;
    imagenes: String[];
    video: string;
    largo: Number;
    alto: Number;
    ancho: Number;
    peso: Number;
}